requirements.txt
torch
git+https://github.com/facebookresearch/detectron2.git
torchvision
torchaudio
streamlit==0.84.0
Pillow==8.3.0
pandas==1.3.0
matplotlib==3.4.2
gdown==3.13.0
numpy
albumentations==1.0.0
opencv-python
tqdm==4.61.1
